requirements.txt
flask
requests
beautifulsoup4
groq
python-dotenv